+++ data/label.json
@@ -8,5 +8,15 @@
     "id": 824,
     "color": "red",
     "title": "light"
+  },
+  {
+    "id": 88,
+    "color": "red",
+    "title": "niv"
+  },
+  {
+    "id": 747,
+    "color": "green",
+    "title": "nice"
   }
 ]
